feat: move to raw u64 for the vm

+++ src-vm/value.zig
@@ -1,6 +1,6 @@
 const std = @import("std");
 
-pub const Value = f64;
+pub const Value = u64;
 
 pub fn print_value(value: Value) void {
     std.debug.print("{d}", .{value});
